Avoid escaping directory structures in sed regex
Changed the sed regex delimiting character from a / to a ?
The script substitutes a string $raspap_dir.  This string can be changed
and it will not be escaped. It is not possible to escape the string
without the use of additional regex. So, the sed regex delimiting
character has been changed to eliminate the need for escaping.