Fix dashboard control issues
Commands to set interface up or down used ifconfig to check if interface
was already up or down. The rest of the page uses "ip s a".  The results
of these two checks sometimes conflict.  As ifconfig is now obsolete
(refer https://linux.die.net/man/8/ifconfig), script has been edited so
that everything uses the "ip s a" check.
Warnings that the interface was already up/down were output in plain
text.  Edited to output these warnings as status->messages.
Refresh button was using javascript to reload the identical page.  This
had the consequence of re-submitting the form as well, including Start
and Stop commands.  Replaced the javascript reload with a link back to
the dashboard page. Used a button class id that has a nifty reload
symbol.
Page was attempting to both change the status of the interface and check
the status of the interface on the same page load.  This is not
possible.  Added a javascript timed page refresh after an interface
state change so that the interface status can be correctly displayed.